improved logic and fixed dates bug

+++ src/app.js
@@ -1,33 +1,54 @@
 const express = require("express");
-const cors = require("cors")
-const bodyParser = require("body-parser");
+const cors = require("cors");
 const { PrismaClient } = require("@prisma/client");
 
 const app = express();
 const prisma = new PrismaClient();
-const port = 3000;
-app.use(cors())
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
+const port = 8000;
+app.use(cors());
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
 app.use((req, res, next) => {
   console.log(`${req.method} request to ${req.url}`);
   next();
 });
 
 app.get("/todo/all", async (req, res) => {
-  const todos = await prisma.todo.findMany();
+  const todos = await prisma.todo.findMany({
+    where: {
+      completed: false,
+    },
+    orderBy: {
+      createdAt: "asc",
+    },
+  });
+  res.json(todos);
+});
+
+app.get("/todo/completed", async (req, res) => {
+  const todos = await prisma.todo.findMany({
+    where: {
+      completed: true,
+    },
+    orderBy: {
+      createdAt: "asc",
+    },
+  });
   res.json(todos);
 });
 
 app.post("/todo/add", async (req, res) => {
   const todo = req.body.todo;
-  if (!todo || !todo.title) {
-    res.status(400).json({ error: "Invalid request body" });
+  if (todo) {
+    // res.status(400).json({ error: "Invalid request body" });
+    // return;
   }
   let newTodo;
   try {
     newTodo = await prisma.todo.create({
-      data: todo,
+      data: {
+        ...todo,
+      },
     });
   } catch (e) {
     console.log(e);
@@ -36,11 +57,41 @@ app.post("/todo/add", async (req, res) => {
   res.json({ newTodo });
 });
 
+app.put("/todo/complete", async (req, res) => {
+  const id = req.body.id;
+  const completed = req.body.completed;
+  if (id === undefined || completed === undefined) {
+    res.status(400).json({ error: "Invalid request body" });
+    return;
+  }
+  console.log(id);
+  try {
+    updatedTodo = await prisma.todo.update({
+      where: {
+        id: Number(id),
+      },
+      data: {
+        completed: completed,
+      },
+    });
+  } catch (e) {
+    console.log(e);
+    res.status(400).json({ error: "Couldn't update todo" });
+  }
+  res.json({ updatedTodo });
+})
+
 app.put("/todo/update", async (req, res) => {
   const todo = req.body.todo;
+  if (!todo) {
+    res.status(400).json({ error: "Invalid request body" });
+    return;
+  }
+  console.log(todo);
   const todoId = todo.id;
   delete todo.id;
   let updatedTodo;
+  // TODO update only changed fields
   try {
     updatedTodo = await prisma.todo.update({
       where: {
@@ -48,6 +99,8 @@ app.put("/todo/update", async (req, res) => {
       },
       data: {
         ...todo,
+        date: todo.date ? new Date(todo.date) : null,
+        deadline: todo.deadline ? new Date(todo.deadline) : null,
       },
     });
   } catch (e) {
@@ -59,6 +112,10 @@ app.put("/todo/update", async (req, res) => {
 
 app.delete("/todo/delete", async (req, res) => {
   const id = req.body.id;
+  if (!id) {
+    res.status(400).json({ error: "Invalid request body" });
+    return;
+  }
   try {
     await prisma.todo.delete({
       where: {
